requirements.txt
Django==3.2.4
djangorestframework
django-crispy-forms
whitenoise
mpmath
numpy
sympy
opencv-python
pytesseract
wordcloud